DOWNGRADED: to Swift 4.2 except enum Result
This commit goes with source code downgraded to Swift 4.2 except enum Result.

Also, with this commit:

UPDATED: Pods.
