+++ ios.native.weather/Weather/Business content/Business logic/Services/UserPreferences.swift
@@ -28,7 +28,7 @@ class UserPreferences {
                 from: settingsData,
                 options: [],
                 format: nil) as? [String: Any],
-              let settingsPreferences = settingsPlist[bundle_PreferencesItems] as? [[String: Any]]
+            let settingsPreferences = settingsPlist?[bundle_PreferencesItems] as? [[String: Any]]
         else { return }
 
         var defaultsToRegister = [String: Any]()
