+++ ios.native.weather/Weather/Business content/Modules/Weather/WeatherViewController.swift
@@ -135,7 +135,7 @@ class WeatherViewController: UIViewController {
         if let choice = isDarkModeSettingsChanged() {
             changeDarkModeManually(choice)
         }
-        print("greetings".localized_value + " ^_^ it's " + DarkMode.Style.description)
+        print("greetings".localized_value + " ^_^ it's " + DarkMode.style.description)
 
         // Start activites
         guard let view = view as? WeatherLayoutViewProtocol else { return }
@@ -162,7 +162,7 @@ class WeatherViewController: UIViewController {
     }
 
     @objc private func makeUp() {
-        self.view.backgroundColor = .systemYellow_Adapted
+        self.view.backgroundColor = .perseusYellow
     }
 
     // MARK: - Other Methods (Not Business Logic Related)
